ai_flask_backend/requirements.txt

#flask
#flask-cors
#ollama

flask
flask-cors
google-genai
